refactored google and bing label implementation into abstract base...
